requirements.txt
fastapi
uvicorn
jinja2
requests
google-generativeai
pydantic
python-dotenv